+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - changed "Switch Materials" existing material detection to search for any material with "Quest" in name
 - changed "Remove Components" component to only show index if more than 1
 - fixed "Remove Blacklisted Components" component not including disabled objects
+- add issue if using an old Unity constraint
 
 # 2.0.0